% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agp.R
\docType{package}
\name{agp}
\alias{agp}
\alias{agp-package}
\title{agp: An R package for agent-based modeling with genetic programming}
\description{
agp: An R package for agent-based modeling with genetic programming
}

